Search refine in progress

+++ resources/views/remittance/search-result.blade.php
@@ -30,8 +30,18 @@
 	      <strong>Name: {{ $searchName }}</strong>
 	    @elseif ($searchBy == 'family')
 	      <strong>Family Code: {{ $searchFamilyCode }}</strong>
+	      @if ($status == 'family_not_found')
+	        <p style="color: red;">
+		  Family {{ $searchFamilyCode}} does not exist.
+	        </p>
+	      @endif
 	    @elseif ($searchBy == 'lot')
 	      <strong>Lot: {{ $searchLotNum }}</strong>
+	      @if ($status == 'lot_not_found')
+	        <p style="color: red;">
+		  Lot {{ $searchLotNum }} does not exist.
+	        </p>
+	      @endif
 	    @elseif ($searchBy == 'serial')
 	      <strong>Serial num: {{ $searchSerialNum }}</strong>
 	    @elseif ($searchBy == 'none')
@@ -53,25 +63,27 @@
 	        </tr>
 	      </thead>
 	      <tbody>
-	        @foreach ($remittances as $remittance)
-		  <tr>
-		    <td>{{  $remittance->remittance_id  }}</td>
-		    <td>
-		    {{ $remittance->family->family_code }}@if($remittance->family->fcode_check_digit !== NULL){{ $remittance->family->fcode_check_digit }}@else N @endif
-                    </td>
-		    <td>
-		      {{ $remittance->submitter->person->first_name }}
-		      {{ $remittance->submitter->person->middle_name }}
-		      {{ $remittance->submitter->person->last_name }}
-		    </td>
-		    <td>{{  $remittance->submitted_date  }}</td>
-		    <td>
-		      <a href="/rmt/{{ $remittance->remittance_id }}">View</a> &nbsp;&nbsp;
-		      <a href="">Update</a> &nbsp;&nbsp;
-		      <a href="/rmt/delete/{{ $remittance->remittance_id }}">Delete</a>
-		    </td>
-		  </tr>
-	        @endforeach
+		@if ($remittances != null)
+	          @foreach ($remittances as $remittance)
+		    <tr>
+		      <td>{{  $remittance->remittance_id  }}</td>
+		      <td>
+		      {{ $remittance->family->family_code }}@if($remittance->family->fcode_check_digit !== NULL){{ $remittance->family->fcode_check_digit }}@else N @endif
+                      </td>
+		      <td>
+		        {{ $remittance->submitter->person->first_name }}
+		        {{ $remittance->submitter->person->middle_name }}
+		        {{ $remittance->submitter->person->last_name }}
+		      </td>
+		      <td>{{  $remittance->submitted_date  }}</td>
+		      <td>
+		        <a href="/rmt/{{ $remittance->remittance_id }}">View</a> &nbsp;&nbsp;
+		        <a href="">Update</a> &nbsp;&nbsp;
+		        <a href="/rmt/delete/{{ $remittance->remittance_id }}">Delete</a>
+		      </td>
+		    </tr>
+	          @endforeach
+		@endif
 	      </tbody>
 	    </table>
 	  </p>
